updated login/signup screen
