+++ src/app/auth/login/actions.ts
@@ -1,0 +1,28 @@
+'use server'
+
+import { revalidatePath } from 'next/cache'
+import { redirect } from 'next/navigation'
+
+import { createClient } from '@/utils/supabase/server'
+
+export async function login(email: string, password: string) {
+    const supabase = createClient()
+
+
+    const data = {
+        email: email,
+        password: password,
+    }
+
+    console.log(data)
+
+    const { error } = await supabase.auth.signInWithPassword(data)
+
+    if (error) {
+        redirect('/error')
+    }
+
+    revalidatePath('/', 'layout')
+    redirect('/')
+}
+
