+++ src/app/auth/login/page.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { Button } from "@/components/ui/button"
 import {
     Card,
@@ -10,37 +12,53 @@ import {
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import Link from "next/link";
+import {login} from "@/app/auth/login/actions";
+import {useState} from "react";
+
 
 const LoginPage = () => {
+
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+
+    const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+        event.preventDefault();
+        login(email, password)
+    };
+
     return (
         <div className="flex items-center justify-center min-h-screen bg-gray-100">
             <Card className="w-full max-w-sm">
-                <CardHeader>
-                    <CardTitle className="text-2xl">Login with your Email</CardTitle>
-                    <CardDescription>
-                        Enter your email below to login to your account.
-                    </CardDescription>
-                </CardHeader>
-                <CardContent className="grid gap-4">
-                    <div className="grid gap-2">
-                        <Label htmlFor="email">Email</Label>
-                        <Input id="email" type="email" placeholder="example@mail.com" required/>
-                    </div>
-                    <div className="grid gap-2">
-                        <Label htmlFor="password">Password</Label>
-                        <Input id="password" type="password" placeholder="Enter your password" required/>
-                    </div>
-                </CardContent>
-                <CardFooter>
-                    <Button className="w-full">Login</Button>
-                </CardFooter>
+                <form onSubmit={handleSubmit}>
+                    <CardHeader>
+                        <CardTitle className="text-2xl">Login with your Email</CardTitle>
+                        <CardDescription>
+                            Enter your email below to login to your account.
+                        </CardDescription>
+                    </CardHeader>
+                    <CardContent className="grid gap-4">
+                        <div className="grid gap-2">
+                            <Label htmlFor="email">Email</Label>
+                            <Input id="email" type="email" value={email} onChange={(e) => setEmail(e.target.value)} placeholder="example@mail.com" required/>
+                        </div>
+                        <div className="grid gap-2">
+                            <Label htmlFor="password">Password</Label>
+                            <Input id="password" type="password" value={password} onChange={(e) => setPassword(e.target.value)} placeholder="Enter your password" required/>
+                        </div>
+                    </CardContent>
+                    <CardFooter>
+                        <Button type="submit" className="w-full" >Login</Button>
+                    </CardFooter>
+                </form>
                 <div className="flex justify-center pb-2">
                     Don’t have an account yet? <Link href="/auth/signup" className="font-bold underline"> Sign up </Link>
                 </div>
+
             </Card>
 
         </div>
     )
 }
+
 
 export default LoginPage
