+++ src/app/auth/signup/actions.ts
@@ -1,0 +1,25 @@
+'use server'
+
+import {createClient} from "@/utils/supabase/server";
+import {redirect} from "next/navigation";
+import {revalidatePath} from "next/cache";
+
+export async function signup(formData: FormData) {
+    const supabase = createClient()
+
+    // type-casting here for convenience
+    // in practice, you should validate your inputs
+    const data = {
+        email: formData.get('email') as string,
+        password: formData.get('password') as string,
+    }
+
+    const { error } = await supabase.auth.signUp(data)
+
+    if (error) {
+        redirect('/error')
+    }
+
+    revalidatePath('/', 'layout')
+    redirect('/')
+}
