+++ src/app/error/page.tsx
@@ -1,0 +1,7 @@
+export default function ErrorPage() {
+    return (
+        <div className="flex items-center justify-center min-h-screen bg-gray-100">
+            <h1>Error Page</h1>
+        </div>
+    )
+}
